+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Version 2.x (unreleased)
 ------------------------
 
-* Added ``DjangoFilterBackend.get_schema_operation_parameters()`` for DRF 3.9+ ``generateschema``.
+* Added ``DjangoFilterBackend.get_schema_operation_parameters()`` for DRF 3.10+ OpenAPI schema generation.
 
 Version 2.1 (2019-1-20)
 -----------------------